+++ repo_monitor/python/tests/data/test_coverage.xml
@@ -1,0 +1,146 @@
+<?xml version="1.0" ?>
+<coverage branch-rate="0" line-rate="0.883" timestamp="1493359236075" version="4.0.3">
+	<!-- Generated by coverage.py: https://coverage.readthedocs.org -->
+	<!-- Based on https://raw.githubusercontent.com/cobertura/web/f0366e5e2cf18f111cbd61fc34ef720a6584ba02/htdocs/xml/coverage-03.dtd -->
+	<sources>
+		<source>/home/michael/projects/scripts/file-kvstore</source>
+		<source>/home/michael/projects/scripts/file-kvstore/file_kvstore</source>
+	</sources>
+	<packages>
+		<package branch-rate="0" complexity="0" line-rate="0.8036" name="file_kvstore">
+			<classes>
+				<class branch-rate="0" complexity="0" filename="file_kvstore/__init__.py" line-rate="1" name="__init__.py">
+					<methods/>
+					<lines/>
+				</class>
+				<class branch-rate="0" complexity="0" filename="file_kvstore/constants.py" line-rate="1" name="constants.py">
+					<methods/>
+					<lines>
+						<line hits="1" number="1"/>
+					</lines>
+				</class>
+				<class branch-rate="0" complexity="0" filename="file_kvstore/io.py" line-rate="0.7083" name="io.py">
+					<methods/>
+					<lines>
+						<line hits="1" number="1"/>
+						<line hits="1" number="4"/>
+						<line hits="1" number="5"/>
+						<line hits="1" number="8"/>
+						<line hits="1" number="9"/>
+						<line hits="1" number="10"/>
+						<line hits="1" number="11"/>
+						<line hits="1" number="14"/>
+						<line hits="1" number="15"/>
+						<line hits="1" number="16"/>
+						<line hits="1" number="17"/>
+						<line hits="1" number="18"/>
+						<line hits="1" number="19"/>
+						<line hits="1" number="20"/>
+						<line hits="1" number="21"/>
+						<line hits="1" number="25"/>
+						<line hits="0" number="26"/>
+						<line hits="0" number="27"/>
+						<line hits="0" number="28"/>
+						<line hits="0" number="29"/>
+						<line hits="0" number="30"/>
+						<line hits="1" number="33"/>
+						<line hits="0" number="34"/>
+						<line hits="0" number="35"/>
+					</lines>
+				</class>
+				<class branch-rate="0" complexity="0" filename="file_kvstore/parser.py" line-rate="0.8889" name="parser.py">
+					<methods/>
+					<lines>
+						<line hits="1" number="1"/>
+						<line hits="1" number="2"/>
+						<line hits="1" number="4"/>
+						<line hits="1" number="5"/>
+						<line hits="1" number="8"/>
+						<line hits="1" number="9"/>
+						<line hits="1" number="10"/>
+						<line hits="1" number="13"/>
+						<line hits="1" number="17"/>
+						<line hits="1" number="20"/>
+						<line hits="1" number="24"/>
+						<line hits="1" number="25"/>
+						<line hits="1" number="26"/>
+						<line hits="1" number="27"/>
+						<line hits="1" number="28"/>
+						<line hits="1" number="31"/>
+						<line hits="0" number="32"/>
+						<line hits="0" number="33"/>
+					</lines>
+				</class>
+				<class branch-rate="0" complexity="0" filename="file_kvstore/utils.py" line-rate="0.8462" name="utils.py">
+					<methods/>
+					<lines>
+						<line hits="1" number="1"/>
+						<line hits="1" number="3"/>
+						<line hits="1" number="4"/>
+						<line hits="1" number="5"/>
+						<line hits="1" number="8"/>
+						<line hits="1" number="9"/>
+						<line hits="1" number="10"/>
+						<line hits="1" number="11"/>
+						<line hits="1" number="12"/>
+						<line hits="1" number="14"/>
+						<line hits="1" number="17"/>
+						<line hits="0" number="18"/>
+						<line hits="0" number="19"/>
+					</lines>
+				</class>
+			</classes>
+		</package>
+		<package branch-rate="0" complexity="0" line-rate="1" name="file_kvstore.tests">
+			<classes>
+				<class branch-rate="0" complexity="0" filename="file_kvstore/tests/__init__.py" line-rate="1" name="__init__.py">
+					<methods/>
+					<lines/>
+				</class>
+				<class branch-rate="0" complexity="0" filename="file_kvstore/tests/test_basic.py" line-rate="1" name="test_basic.py">
+					<methods/>
+					<lines>
+						<line hits="1" number="1"/>
+						<line hits="1" number="2"/>
+						<line hits="1" number="4"/>
+						<line hits="1" number="7"/>
+						<line hits="1" number="8"/>
+						<line hits="1" number="9"/>
+						<line hits="1" number="10"/>
+						<line hits="1" number="11"/>
+						<line hits="1" number="12"/>
+						<line hits="1" number="13"/>
+						<line hits="1" number="14"/>
+						<line hits="1" number="16"/>
+						<line hits="1" number="17"/>
+						<line hits="1" number="19"/>
+						<line hits="1" number="20"/>
+						<line hits="1" number="21"/>
+						<line hits="1" number="22"/>
+						<line hits="1" number="24"/>
+						<line hits="1" number="25"/>
+						<line hits="1" number="26"/>
+						<line hits="1" number="27"/>
+						<line hits="1" number="31"/>
+						<line hits="1" number="32"/>
+						<line hits="1" number="33"/>
+						<line hits="1" number="34"/>
+						<line hits="1" number="38"/>
+						<line hits="1" number="39"/>
+						<line hits="1" number="40"/>
+						<line hits="1" number="41"/>
+						<line hits="1" number="45"/>
+						<line hits="1" number="46"/>
+						<line hits="1" number="47"/>
+						<line hits="1" number="48"/>
+						<line hits="1" number="52"/>
+						<line hits="1" number="53"/>
+						<line hits="1" number="54"/>
+						<line hits="1" number="55"/>
+						<line hits="1" number="56"/>
+					</lines>
+				</class>
+			</classes>
+		</package>
+	</packages>
+</coverage>
